requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
pydantic==2.5.0

# Web Scraping & Data Processing
requests==2.31.0
beautifulsoup4==4.12.2
scrapy==2.11.0
newspaper3k==0.2.8
python-dateutil==2.8.2

# Vector Database & Search
faiss-cpu==1.7.4
chromadb==0.4.15
sentence-transformers==2.2.2
langchain==0.0.335
langchain-community==0.0.5

# Language Processing
transformers==4.35.2
torch==2.1.1
langdetect==1.0.9
fasttext==0.9.2

# Translation
sentencepiece==0.1.99
sacremoses==0.0.53

# LLM Integration
ollama==0.1.8
openai==1.3.8

# Voice Processing
speechrecognition==3.10.0
pydub==0.25.1
soundfile==0.12.1
librosa==0.10.1

# Text-to-Speech
TTS==0.19.0
gTTS==2.4.0
pygame==2.5.2

# Database
sqlite3
sqlalchemy==2.0.23
alembic==1.12.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
tqdm==4.66.1
schedule==1.2.0

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.5.0

# Docker & Deployment
gunicorn==21.2.0
